+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - **FIX**: [metro rate not updated after `INIT.SCENE`](https://github.com/monome/teletype/issues/174)
 - **NEW**: new ops: SEED, R.SEED, TOSS.SEED, DRUNK.SEED, P.SEED, PROB.SEED
 - **NEW**: new op: SCENE.G
+- **NEW**: new op: SCRIPT.POL, alias $.POL
 - **NEW**: new ansible remote ops: ANS.G, ANS.G.P, ANS.G.LED, ANS.A, ANS.A.LED
 - **NEW**: new kria remote ops: KR.CUE, KR.PG